delete file and moved the tests for dedicated files
